Changed implementation

+++ src/main/java/com/example/TravelPlanner/travelplanning/services/EventServiceImpl.java
@@ -1,15 +1,18 @@
 package com.example.TravelPlanner.travelplanning.services;
 
 import com.example.TravelPlanner.common.exceptions.custom.BadRequest;
+import com.example.TravelPlanner.common.exceptions.custom.NoPermissionException;
 import com.example.TravelPlanner.common.exceptions.custom.OverlappingEventsException;
 import com.example.TravelPlanner.common.exceptions.custom.entitynotfound.EventNotFoundException;
 import com.example.TravelPlanner.common.utils.CentralSupport;
 import com.example.TravelPlanner.common.utils.mappers.event.EventMapper;
+import com.example.TravelPlanner.travelplanning.common.CheckService;
 import com.example.TravelPlanner.travelplanning.common.enums.PlaceStatus;
 import com.example.TravelPlanner.travelplanning.dto.event.EventCreateDTO;
 import com.example.TravelPlanner.travelplanning.dto.event.EventDTO;
 import com.example.TravelPlanner.travelplanning.entities.Event;
 import lombok.RequiredArgsConstructor;
+import org.hibernate.annotations.Check;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
@@ -27,6 +30,7 @@ public class EventServiceImpl implements EventService{
 
     private final CentralSupport centralSupport;
     private final EventMapper eventMapper;
+    private final CheckService checkService;
 
     @Override
     public List<EventDTO> listAllEventsByTravelPlan(Long planId) {
@@ -35,11 +39,7 @@ public class EventServiceImpl implements EventService{
 
     @Override
     public EventDTO getEventById(Long eventId) {
-        Optional<Event> optionalEvent = centralSupport.getEventRepository().findById(eventId);
-        if (optionalEvent.isEmpty()){
-            throw new EventNotFoundException(eventId);
-        }
-        return eventMapper.mapEventToEventDTO(optionalEvent.get());
+        return eventMapper.mapEventToEventDTO(checkService.checkEventExistence(eventId));
     }
 
     @Override
@@ -57,51 +57,50 @@ public class EventServiceImpl implements EventService{
         ) {
             throw new OverlappingEventsException();
 
-        } else {
-            Event newEvent = eventMapper.mapEventCreateDTOtoEvent(eventCreateDTO, travelPlanId, userId);
-            newEvent = centralSupport.getEventRepository().save(newEvent);
-            return eventMapper.mapEventToEventDTO(newEvent);
         }
-
+        Event newEvent = eventMapper.mapEventCreateDTOtoEvent(eventCreateDTO, travelPlanId, userId);
+        newEvent = centralSupport.getEventRepository().save(newEvent);
+        return eventMapper.mapEventToEventDTO(newEvent);
     }
 
     @Transactional
     @Override
-    @PreAuthorize("@permissionSecurityService.hasEditorPermission(#travelPlanId, #userId)")
     public void updateEvent(EventDTO eventDTO, Long travelPlanId, UUID userId) {
-//        Optional<Event> eventOptional = mappingSupport.getEventRepository().findById(eventDTO.getId());
-//        if(eventOptional.isEmpty()){
-//            throw new EventNotFoundException(eventDTO.getId());
-//        }
-//        Event event = eventOptional.get();
-        Event event = centralSupport.getEventRepository().findById(eventDTO.getId()).get();
-        if(eventDTO.getPlaceStatus() == PlaceStatus.CONCRETE &&
-                (event.getPlaceStatus() == PlaceStatus.SUGGESTED || event.getPlaceStatus() == PlaceStatus.VOTING)) {
-            centralSupport.getEventRepository().updateEventStatusByTravelPlanAndTime(
-                    travelPlanId,
-                    event.getPlaceStatus(),
-                    event.getPlaceStatus(),
-                    event.getStartTime(),
-                    event.getEndTime()
-            );
-            event.setPlaceStatus(PlaceStatus.CONCRETE);
+        Event event = checkService.checkEventExistence(eventDTO.getId());
+        if(event.getCreator().getId() == userId || event.getTravelPlan().getOwner().getId() == userId){
+            if(eventDTO.getPlaceStatus() == PlaceStatus.CONCRETE &&
+                    (event.getPlaceStatus() == PlaceStatus.SUGGESTED || event.getPlaceStatus() == PlaceStatus.VOTING)) {
+                centralSupport.getEventRepository().updateEventStatusByTravelPlanAndTime(
+                        travelPlanId,
+                        event.getPlaceStatus(),
+                        event.getPlaceStatus(),
+                        event.getStartTime(),
+                        event.getEndTime()
+                );
+                event.setPlaceStatus(PlaceStatus.CONCRETE);
 
-        } else {
-            event.setPlaceStatus(PlaceStatus.SUGGESTED);
+            } else {
+                event.setPlaceStatus(PlaceStatus.SUGGESTED);
+            }
+            event.setLocation(event.getLocation());
+            event.setTitle(eventDTO.getTitle());
+            event.setDescription(eventDTO.getDescription());
+            event.setStartTime(eventDTO.getStartTime());
+            event.setEndTime(eventDTO.getEndTime());
+            centralSupport.getEventRepository().save(event);
         }
 
-        eventMapper.setEventLocation(event, eventDTO.getLoc());
-        event.setTitle(eventDTO.getTitle());
-        event.setDescription(eventDTO.getDescription());
-        event.setStartTime(eventDTO.getStartTime());
-        event.setEndTime(eventDTO.getEndTime());
-        centralSupport.getEventRepository().save(event);
     }
 
     @Override
     @Transactional
-    public void deleteEvent(Long eventId) {
-        centralSupport.getEventRepository().deleteById(eventId);
+    public void deleteEvent(Long eventId, UUID userId) {
+        Event event = checkService.checkEventExistence(eventId);
+        if (event.getCreator().getId() == userId || event.getTravelPlan().getOwner().getId() == userId) {
+            centralSupport.getEventRepository().deleteById(eventId);
+        } else {
+            throw new NoPermissionException("Not permitted to perform event deletion!");
+        }
     }
 
 }
